manually implemented KNN

+++ app/src/main/java/com/cs125/helth/Run.java
@@ -3,9 +3,9 @@ package com.cs125.helth;
 public class Run {
     int a_id;
     int u_id;
-    float total_time;
+    double total_time;
     int average_heart_rate;
-    float pace;
-    float total_distance;
+    double pace;
+    double total_distance;
     String date;
 }
